latihan pertemuan 1 latifha

+++ pertemuan1/casting.py
@@ -12,7 +12,7 @@ y = int(5.8) # y akan jadi 5
 z = int("310") # z akan jadi 310
 
 #float()
-x = float(1)     # x jadi 1.0
+x = float(25)     # x jadi 25.0
 y = float(2.8)   # y jadi 2.8
 z = float("3")   # z jadi 3.0
